# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClickDeal Assistant API",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "https://api.clickdeal.site"
    }
  ],
  "paths": {
    "/api/stock/{variantId}": {
      "get": {
        "summary": "Live stock for a Shopify variant",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "variantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "variantId": {
                      "type": "string"
                    },
                    "quantity": {
                      "type": "integer"
                    },
                    "tracked": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/price/{handle}": {
      "get": {
        "summary": "Live price for first variant by product handle",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "productHandle": {
                      "type": "string"
                    },
                    "productTitle": {
                      "type": "string"
                    },
                    "variantId": {
                      "type": "string"
                    },
                    "variantTitle": {
                      "type": "string"
                    },
                    "price": {
                      "type": "object",
                      "properties": {
                        "amount": {
                          "type": "string"
                        },
                        "currencyCode": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/orders": {
      "post": {
        "summary": "Create Shopify Draft Order and WhatsApp notify",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "productHandle": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer"
                  }
                },
                "required": [
                  "name",
                  "phone",
                  "address",
                  "city",
                  "productHandle",
                  "quantity"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "draftOrderId": {
                      "type": "string"
                    },
                    "invoiceUrl": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    }
  }
}